refactor: consolidate atomic-file-write helpers
Three independent atomic-write helpers had drifted across the codebase.
Consolidate them into a single canonical helper in atomic_write.go.

- Move atomicWriteFile from daemon.go to a dedicated atomic_write.go
  (no behavior change, better location — it's used by review files,
  config, sessions, plans, and aider integration).
- Delete writeFileMkdirAtomic in aider_install.go. Replace its three
  call sites with atomicWriteFile(path, data, 0o644). The old helper
  hardcoded perm via O_CREATE flags and skipped Chmod; canonical helper
  now guarantees 0o644 regardless of umask.
- Rewrite plans.savePlanSessions to use atomicWriteFile instead of
  inline os.WriteFile + os.Rename. This adds the missing fsync — a real
  durability bug. Without fsync between rename and the kernel flushing
  the data block, a power loss left a zero-byte file at the target
  (the inode rename committed but the data did not). The advisory
  flock around the function is preserved.
- Drop redundant os.MkdirAll in saveCritJSON; atomicWriteFile already
  creates the parent directory.
