+++ daemon.go
@@ -108,44 +108,6 @@ func reviewFilePath(key string) (string, error) {
 	return filepath.Join(dir, key+".json"), nil
 }
 
-// atomicWriteFile writes data to the target path atomically using temp file + fsync + rename.
-func atomicWriteFile(target string, data []byte, perm os.FileMode) error {
-	dir := filepath.Dir(target)
-	if err := os.MkdirAll(dir, 0700); err != nil {
-		return fmt.Errorf("creating directory %s: %w", dir, err)
-	}
-
-	tmp, err := os.CreateTemp(dir, filepath.Base(target)+"*.tmp")
-	if err != nil {
-		return fmt.Errorf("create temp file: %w", err)
-	}
-	tmpName := tmp.Name()
-
-	if _, err := tmp.Write(data); err != nil {
-		tmp.Close()
-		os.Remove(tmpName)
-		return fmt.Errorf("write temp file: %w", err)
-	}
-	if err := tmp.Sync(); err != nil {
-		tmp.Close()
-		os.Remove(tmpName)
-		return fmt.Errorf("sync temp file: %w", err)
-	}
-	if err := tmp.Close(); err != nil {
-		os.Remove(tmpName)
-		return fmt.Errorf("close temp file: %w", err)
-	}
-	if err := os.Chmod(tmpName, perm); err != nil {
-		os.Remove(tmpName)
-		return fmt.Errorf("chmod temp file: %w", err)
-	}
-	if err := os.Rename(tmpName, target); err != nil {
-		os.Remove(tmpName)
-		return fmt.Errorf("rename temp file to target: %w", err)
-	}
-	return nil
-}
-
 // writeSessionFile writes a session entry to ~/.crit/sessions/<key>.json.
 func writeSessionFile(key string, entry sessionEntry) error {
 	path, err := sessionFilePath(key)
