+++ plans.go
@@ -259,14 +259,12 @@ func savePlanSlug(sessionID, slug string) error {
 		return err
 	}
 
-	// Atomic write via temp file + rename to prevent partial writes.
-	tmpPath := path + ".tmp"
-	if err := os.WriteFile(tmpPath, out, 0644); err != nil {
-		return fmt.Errorf("writing plan sessions temp file: %w", err)
-	}
-	if err := os.Rename(tmpPath, path); err != nil {
-		os.Remove(tmpPath)
-		return fmt.Errorf("renaming plan sessions temp file: %w", err)
+	// Route through atomicWriteFile (atomic_write.go) so we get fsync
+	// before rename. Without fsync, a power loss between rename and the
+	// kernel flushing the data block leaves a zero-byte file on disk
+	// (the inode is renamed but the data was never written).
+	if err := atomicWriteFile(path, out, 0o644); err != nil {
+		return fmt.Errorf("writing plan sessions: %w", err)
 	}
 	return nil
 }
